+++ cursos/avaliacao-de-feridas-e-curativos.php
@@ -1,0 +1,6 @@
+<?php
+require __DIR__ . '/../includes/functions.php';
+require __DIR__ . '/../includes/env.php';
+require __DIR__ . '/../includes/turmas.php';
+require __DIR__ . '/../includes/curso-page.php';
+render_curso_page('avaliacao-de-feridas-e-curativos');